+++ app/frontend/javascripts/components/ConditionDiseaseSearch/ConditionDiseaseSearchOntologyCard.js
